gir-caddie-mvp/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==2.1.2
openai>=1.40.0
tenacity>=8.2.3
python-dotenv==1.0.1
